+++ Sources/LiveKit/Track/Track.swift
@@ -344,20 +344,3 @@ extension Track {
         delegates.notify(label: label, fnc)
     }
 }
-
-// MARK: - Objective-C Support
-
-extension Track {
-
-    @objc(start)
-    @discardableResult
-    public func startObjC() -> Promise<Bool>.ObjCPromise<NSNumber> {
-        start().asObjCPromise()
-    }
-
-    @objc(stop)
-    @discardableResult
-    public func stopObjC() -> Promise<Bool>.ObjCPromise<NSNumber> {
-        stop().asObjCPromise()
-    }
-}
